% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas.R
\name{betas}
\alias{betas}
\title{Extract betas values of a model for specific observations}
\usage{
betas(object, newdata, ...)
}
\arguments{
\item{object}{a model}

\item{newdata}{new observation(s) with columns that correspond to variables used in the model}

\item{...}{unused additional parameters}
}
\description{
Extract betas values of a model for specific observations
}
\author{
Joseph Larmarange
}
