% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_interactions.R
\name{break_down}
\alias{break_down}
\title{Model Agnostic Experimental Approach to Break Down Plots with Interactions}
\usage{
break_down(explainer, new_observation, check_interactions = TRUE,
  keep_distributions = FALSE)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by function `DALEX::explain()`.}

\item{new_observation}{a new observation with columns that corresponds to variables used in the model}

\item{check_interactions}{the orgin/baseline for the breakDown plots, where the rectangles start. It may be a number or a character "Intercept". In the latter case the orgin will be set to model intercept.}

\item{keep_distributions}{if TRUE, then the distribution of partial predictions is stored in addition to the average.}
}
\value{
an object of the broken class
}
\description{
This function implements decomposition of model predictions with identification
of interactions.
The complexity of this function is O(2*p) for additive models and O(2*p^2) for interactions
}
\examples{
\dontrun{
library("breakDown")
library("randomForest")
library("ggplot2")
set.seed(1313)
model <- randomForest(factor(left)~., data = HR_data, family = "binomial", maxnodes = 5)
predict.function <- function(model, new_observation)
      predict(model, new_observation, type="prob")[,2]
predict.function(model, HR_data[11,-7])
explain_1 <- broken(model, HR_data[11,-7], data = HR_data[,-7],
predict.function = predict.function, direction = "down")
explain_1
plot(explain_1) + ggtitle("breakDown plot (direction=down) for randomForest model")

explain_2 <- broken(model, HR_data[11,-7], data = HR_data[,-7],
predict.function = predict.function, direction = "down", keep_distributions = TRUE)
plot(explain_2, plot_distributions = TRUE) +
         ggtitle("breakDown distributions (direction=down) for randomForest model")

explain_3 <- broken(model, HR_data[11,-7], data = HR_data[,-7],
predict.function = predict.function, direction = "up", keep_distributions = TRUE)
plot(explain_3, plot_distributions = TRUE) +
         ggtitle("breakDown distributions (direction=up) for randomForest model")
}
}
