% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.broken}
\alias{print.broken}
\title{Break Down Print}
\usage{
\method{print}{broken}(x, ...)
}
\arguments{
\item{x}{the model model of 'broken' class}

\item{...}{other parameters}
}
\value{
a data frame
}
\description{
Break Down Print
}
