% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.R
\name{broken}
\alias{broken}
\title{Create the broken object}
\usage{
broken(model, new_observation, ...)
}
\arguments{
\item{model}{a model}

\item{new_observation}{a new observation with collumns that corresponds to variables used in the model}

\item{...}{other parameters}
}
\description{
Create the broken object
}
