% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create.broken}
\alias{create.broken}
\title{Clean the object of the broken class
Internal function}
\usage{
create.broken(broken_intercept, baseline = 0)
}
\value{
enriched broken class
}
\description{
Clean the object of the broken class
Internal function
}
