% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create.broken}
\alias{create.broken}
\title{Clean the object of the broken class
Internatl function}
\usage{
create.broken(broken_intercept)
}
\value{
enriched broken class
}
\description{
Clean the object of the broken class
Internatl function
}
