% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.broken}
\alias{plot.broken}
\title{Break Down PLot}
\usage{
\method{plot}{broken}(x, trans = I, ..., add_contributions = TRUE,
  vcolors = c(`-1` = "#d8b365", `0` = "#f5f5f5", `1` = "#5ab4ac", X =
  "darkgrey"), digits = 3, rounding_function = round)
}
\arguments{
\item{x}{the model model of 'broken' class}

\item{trans}{transformation that shal be applied to scores}

\item{...}{other parameters}

\item{add_contributions}{shall variable contributions to be added on plot?}

\item{vcolors}{named vector with colors}

\item{digits}{number of decimal places (round) or significant digits (signif) to be used.
See the \code{rounding_function} argument}

\item{rounding_function}{function that is to used for rounding numbers.
It may be \code{signif()} which keeps a specified number of significant digits.
Or the default \code{round()} to have the same precision for all components}
}
\value{
a ggplot2 object
}
\description{
Break Down PLot
}
