% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.broken}
\alias{plot.broken}
\title{Break Down PLot}
\usage{
\method{plot}{broken}(x, ...)
}
\arguments{
\item{x}{the model model of 'broken' class}

\item{...}{other parameters}
}
\value{
a ggplot2 object
}
\description{
Break Down PLot
}
