% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.broken}
\alias{print.broken}
\title{Break Down Print}
\usage{
\method{print}{broken}(x, ..., digits = 3, rounding_function = round)
}
\arguments{
\item{x}{the model model of 'broken' class}

\item{...}{other parameters}

\item{digits}{number of decimal places (round) or significant digits (signif) to be used.
See the \code{rounding_function} argument}

\item{rounding_function}{function that is to used for rounding numbers.
It may be \code{signif()} which keeps a specified number of significant digits.
Or the default \code{round()} to have the same precision for all components}
}
\value{
a data frame
}
\description{
Break Down Print
}
